Import React explicitly to each component
This is a config change that was to do with moving to a different
development machine. I think on my old machine I has something
extra setup which would transform React a bit easier without
the need for an explicit import, but not on my new one. I think it
helps to have it imported anyway so I've added it to each file.