Move message dismissal back to the provider
In a previous change, the logic for dismissing a notification
was placed on the message itself, and it would dismiss itself
via a function passed in as a prop. This design was changed here
to leave control in the provider.

Also, remove the button that was placed on the notification that
was used for testing