Theme updates to support color variants
Previously, the basic colors for success, info, warning, and danger
were just stored in the theme as strings, and the `light` variants
were also just strings on the `colors` object in the `theme`. This made it
difficult to apply variations to the colors based on the variant used.

Now, the mentioned colors are now objects which have base, light, and dark
properties which makes it easier to apply subtle vairations.