Add phoshor-icons and style notifications
Add phosphor-icons as the icon library for the project and create
a component for icons in the notification body. With that done,
clean up the content of the notification and the styling so that
it looks as expected. Remove the test button and call the notification
when a service is created or updated.