Notifications (#14)
* Adjust modal height issue

* Create a notification provider to handle messages

* Create an unstyled message component to show in the UI

* Wrap the app in the notification provider so that children can use the notification hook

* Add a test button to summon the notification

* Add duration prop for notification, with 3000ms default

* Rearrange types

* Add basic styles and animations for notifications

* Pause timeout to allow for easier style debugging

* Change animation style for notification

* Add support for multiple messages

* Working on closing notifications

* Import React explicitly to each component

This is a config change that was to do with moving to a different
development machine. I think on my old machine I has something
extra setup which would transform React a bit easier without
the need for an explicit import, but not on my new one. I think it
helps to have it imported anyway so I've added it to each file.

* Move message dismissal back to the provider

In a previous change, the logic for dismissing a notification
was placed on the message itself, and it would dismiss itself
via a function passed in as a prop. This design was changed here
to leave control in the provider.

Also, remove the button that was placed on the notification that
was used for testing

* Theme updates to support color variants

Previously, the basic colors for success, info, warning, and danger
were just stored in the theme as strings, and the `light` variants
were also just strings on the `colors` object in the `theme`. This made it
difficult to apply variations to the colors based on the variant used.

Now, the mentioned colors are now objects which have base, light, and dark
properties which makes it easier to apply subtle vairations.

* Add phoshor-icons and style notifications

Add phosphor-icons as the icon library for the project and create
a component for icons in the notification body. With that done,
clean up the content of the notification and the styling so that
it looks as expected. Remove the test button and call the notification
when a service is created or updated.